created TimelineSimple component
